pattern_match basic page split
